Move NotifyFramebuffer and LoadClut to common code.
