+++ GPU/Common/TextureCacheCommon.h
@@ -26,13 +26,25 @@ enum TextureFiltering {
 	TEX_FILTER_LINEAR_VIDEO = 4,
 };
 
+enum FramebufferNotification {
+	NOTIFY_FB_CREATED,
+	NOTIFY_FB_UPDATED,
+	NOTIFY_FB_DESTROYED,
+};
+
 struct VirtualFramebuffer;
 
 class TextureCacheCommon {
 public:
+	TextureCacheCommon();
 	virtual ~TextureCacheCommon();
 
+	void LoadClut(u32 clutAddr, u32 loadBytes);
+
 	virtual bool SetOffsetTexture(u32 offset);
+
+	// FramebufferManager keeps TextureCache updated about what regions of memory are being rendered to.
+	void NotifyFramebuffer(u32 address, VirtualFramebuffer *framebuffer, FramebufferNotification msg);
 
 	int AttachedDrawingHeight();
 
@@ -115,9 +127,26 @@ public:
 	};
 
 protected:
+	// Can't be unordered_map, we use lower_bound ... although for some reason that compiles on MSVC.
+	typedef std::map<u64, TexCacheEntry> TexCache;
+
 	void GetSamplingParams(int &minFilt, int &magFilt, bool &sClamp, bool &tClamp, float &lodBias, u8 maxLevel);
 
+	virtual bool AttachFramebuffer(TexCacheEntry *entry, u32 address, VirtualFramebuffer *framebuffer, u32 texaddrOffset = 0) = 0;
+	virtual void DetachFramebuffer(TexCacheEntry *entry, u32 address, VirtualFramebuffer *framebuffer) = 0;
+
+	TexCache cache;
+	std::vector<VirtualFramebuffer *> fbCache_;
+
 	TexCacheEntry *nextTexture_;
+
+	// Raw is where we keep the original bytes.  Converted is where we swap colors if necessary.
+	u32 *clutBufRaw_;
+	u32 *clutBufConverted_;
+	u32 clutLastFormat_;
+	u32 clutTotalBytes_;
+	u32 clutMaxBytes_;
+	u32 clutRenderAddress_;
 };
 
 inline bool TextureCacheCommon::TexCacheEntry::Matches(u16 dim2, u8 format2, u8 maxLevel2) {
